Add "Show Rollover with Budget" feature
If my goal is to budget $100/month in a category, and I only spent $70 last month, I need a quick way to see that I only need to budget $70 in the category this month to hit my target. Goals can't do this, and using the "available" column only works until I have a transaction in that category for the month. This way, I can more easily see how much money I had available at the start of the month -- it's the amount budgeted, plus last month's overflow.

I based this feature on display-upcoming-amount.